Drop spurious extsh in barrelgener Obj_SmoothTurnAnglesTowardVelocity
Use (s16) cast instead of (int) on fctiwz results stored to s16 angle
fields (recipe #41/#83c). 96.02 -> 97.22.

+++ src/main/dll/dll_0282_barrelgener.c
@@ -403,8 +403,8 @@ void Obj_SmoothTurnAnglesTowardVelocity(int a, int b, int c, f32 d, f32 e)
 
     if (d != lbl_803E6C38)
     {
-        *(s16*)(a + 4) = (int)(lbl_803E6C98 * (f32) * (s16*)(a + 4));
-        *(s16*)(a + 4) = (int)(oneOverTimeDelta * (lbl_803E6C5C * (clamped * d)) + (f32) * (s16*)(a + 4));
+        *(s16*)(a + 4) = (s16)(lbl_803E6C98 * (f32) * (s16*)(a + 4));
+        *(s16*)(a + 4) = (s16)(oneOverTimeDelta * (lbl_803E6C5C * (clamped * d)) + (f32) * (s16*)(a + 4));
         tmp = *(s16*)(a + 4);
         if (tmp < -0x2000)
         {
